+++ src/pages/chat/[id]/Config/Topic/DefaultContent.tsx
@@ -1,0 +1,29 @@
+import { Icon } from '@lobehub/ui';
+import { Tag, Typography } from 'antd';
+import { useTheme } from 'antd-style';
+import { MessageSquareDashed } from 'lucide-react';
+import { memo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Flexbox } from 'react-layout-kit';
+
+const { Paragraph } = Typography;
+
+const DefaultContent = memo(() => {
+  const { t } = useTranslation('common');
+
+  const theme = useTheme();
+
+  return (
+    <Flexbox align={'center'} gap={8} horizontal>
+      <Flexbox align={'center'} height={24} justify={'center'} width={24}>
+        <Icon color={theme.colorTextDescription} icon={MessageSquareDashed} />
+      </Flexbox>
+      <Paragraph ellipsis={{ rows: 1 }} style={{ margin: 0 }}>
+        {t('topic.defaultTitle')}
+      </Paragraph>
+      <Tag>{t('temp')}</Tag>
+    </Flexbox>
+  );
+});
+
+export default DefaultContent;
